2012.03.22
1. 2) analyse how to make 5 minutes retry interval configurable

Solution - the most convenient way for this - extend it in the same place where cache is configured.
But cache configuration is not kept in cache. 
Solution - extend current implementations of ICache to keep passed configuration and use it in CrlCache class with default value 5 minutes.

2. 5) implement configuration of _maxEqualKeySize for MultipleQuantityCache

Done. Configuration option is "maxEqualKey"

3. 4) replace current QuantityCache implementation with subclass of MultipleQuantityCache

Implemented LeastFrequentlyUsedQuantityCache and LeastRecentlyUsedQuantityCache(similar to QuantityCache).

a) less blocking
b) removes from cache less used (not latest used) value

4. Closed 918 - added description why.

What to do:

1) Finally! unit tests in points 6,7,8
2) Analyse connection and read timeouts during call to CRL server (not defined now at all)
3) Replace current QuantityCache implementation with LeastFrequentlyUsedQuantityCache
4) Add usage of config option, defined in 1. in CrlCache