4. A new class dk.gov.oiosi.Version is added with next java doc comments:
Administrative class to keep track of the version number of the OIORASP release. 
This class is created to be similar to Version classes in Apache project, e.g. org.apache.xalan.Version. 
In future it can be used by configuration procedures to check proper library related environment update. 
Do not mix it with dk.gov.oiosi.common.Version, which is Version of server and contains another information.

It returns version 1.3.0

Unit test: VersionTest