35. CrlCache - IMPORTANT: fixed serious bug with using URL as a key for hash map. Added  CrlCacheDesignIssueAnalysis with reproduce case and explanaitions.
36. Added CrlCacheDesignIssueAnalysis class to list of tests, but with only main method:

This program shows, that it is not really correct to use URL object as a key in collections, because its hash code function uses resolved IP address as a part of hash. 
In OIORASP 1.2.3.HotFix1 at least in the class CrlCache revocation certificates are stored in Hashtable: 
private Hashtable table 
This test shows, that it is not really good, because of the server experiences network problems, it will create a new record in this set! 
Please find example of the program output, generated on Windows 7 OS and Java HotSpot(TM) 64-Bit Server VM (build 20.0-b11, mixed mode), JRE build 1.6.0_25-b06, in file CrlCacheDesignIssueAnalysis_Result.txt